wip. init works, except save config is wrong
